Move experiment-specific files to dedicated folder
