+++ factchecker/experiments/advocate_mediator_climatefeedback/EXPERIMENT_DETAILS.md
@@ -1,0 +1,12 @@
+
+# Advocate_Mediator_Climatefeedback
+
+Based on this paper: ...
+Based on this dataset: ...
+Description of Factchecking Pipeline: ...
+
+How to configure: ...
+How to run: ...
+How to inspect results: ...
+
+...
